requirements.txt
alembic==0.9.8
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.1.1
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
futures==3.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
PyMySQL3==0.5
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.3
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
